Web settings firstLogin cleanup #3082 (#3091)
