+++ configuration-client/src/components/Login/Login.tsx
@@ -5,10 +5,12 @@ import { useContext } from 'react';
 import { User, Lock } from 'tabler-icons-react';
 
 import I18nContext from '../../contexts/i18n-context';
-import { login } from '../../services/auth.service';
+import { create, login } from '../../services/auth.service';
+import SessionContext from '../../contexts/session-context';
 
 const Login = () => {
   const i18n = useContext(I18nContext);
+  const session = useContext(SessionContext);
 
   const form = useForm({
     initialValues: {
@@ -20,10 +22,7 @@ const Login = () => {
   const handleLogin = (values: typeof form.values) => {
     const { username, password } = values;
     login(username, password).then(
-      ({firstLogin}) => {
-        if (firstLogin === "true") {
-          return window.location.href = '/changepassword'
-        }
+      () => {
         window.location.reload();
       },
       (error) => {
@@ -38,10 +37,29 @@ const Login = () => {
     );
   };
 
+  const handleUserCreation = (values: typeof form.values) => {
+    const { username, password } = values;
+    create(username, password).then(
+      () => {
+        window.location.reload();
+      },
+      (error) => {
+        showNotification({
+          id: 'user-creation-error',
+          color: 'red',
+          title: 'Error',
+          message: 'Error on user creation',
+          autoClose: 3000,
+        });
+      }
+    );
+  };
+
   return (
     <Box sx={{ maxWidth: 300 }} mx='auto'>
-      <form onSubmit={form.onSubmit(handleLogin)}>
-        <Text size="xl">Log in</Text>
+      <form onSubmit={form.onSubmit(session.noAdminFound ? handleUserCreation : handleLogin)}>
+          <Text size="xl">Universal Media Server</Text>
+          <Text size="lg">{ session.noAdminFound ? 'Create your first admin user' : 'Log in' }</Text>
         <Space h="md" />
         <TextInput
           required
